GTK dlopen codepath now unloads GTK and GLIB after every API call.
I was looking at VmSize via cat /proc/$pid/status.
At start of the test_opendialog.c, it was 8284kb.
After calling IsAvailable which loads the libraries, it was 107032 kB.
After unloading the libraries, it went back down to 12660 kB.

The memory difference seems significant enough to want to enable this by
default.