Discovered on some distros (SteamOS, Rasbpian Stretch), that the second time opening a panel leads to a crash. Changed code to use RTLD_NODELETE to workaround this problem.
I don't understand why this crash happens (and only why some
platforms...could it be GTK2 vs. GTK3?), but I speculate maybe there is
a static/global variable in GTK that doesn't get reinitialized after
first use, which leads to a stale/dangling pointer and thus a crash.

The RTLD_NODELETE is the easiest way to fix the code without rewriting
all the dlclose stuff.

For now, the dlclose still happens for the NFD_IsAvailable check. This
use case of immediately closing without using seems to avoid any
crashes. It kind of sucks to give up this much RAM just for an
availabilty check, so returning it for this case seems desirable.